Move library media enrichment out of routes
