+++ backend/src/api/libraryRoutes.ts
@@ -1,5 +1,4 @@
 import fs from "node:fs/promises";
-import path from "node:path";
 
 import type { NextFunction, Request, Response } from "express";
 import { Router } from "express";
@@ -11,11 +10,19 @@ import {
   readTrackMetadataOverrides
 } from "../services/indexer/metadataOverrideStore";
 import { IndexStore } from "../services/indexer/indexStore";
+import {
+  attachArtistPhotos,
+  attachCollaborativeAlbumCovers,
+  compareAlbumTrackOrder,
+  getTrackArtistName,
+  normalizeAlbumName,
+  parseCollaborativeAlbumId,
+  resolveAlbumMetadataForTrack,
+  type ResolvedAlbumMetadata
+} from "../services/library/libraryMediaResolver";
 import { filterPlayablePlaylists } from "../services/playlists/playlistStore";
 import { deleteTrackCover } from "../services/storage/coverService";
 import { resolveTrackAbsolutePath } from "../services/storage/storageService";
-import { fileExists, readJsonFile } from "../utils/fs";
-import { resolveDataRelativePath } from "../utils/paths";
 import {
   filterTracks,
   getAdjacentTrack,
@@ -34,8 +41,6 @@ import type { Track } from "../types/library";
 const DEFAULT_TRACKS_PAGE = 1;
 const DEFAULT_TRACKS_LIMIT = 100;
 const MAX_TRACKS_LIMIT = 500;
-const ARTIST_METADATA_FILE = "artist.json";
-const ALBUM_METADATA_FILE = "album.json";
 
 const SUPPORTED_TRACK_SORT_FIELDS = new Set<TrackSortBy>([
   "title",
@@ -256,264 +261,6 @@ function mapTrackOwner(track: Track, ownerNamesById: Map<string, string>): Track
 
 function mapTrackOwners(tracks: Track[], ownerNamesById: Map<string, string>): Track[] {
   return tracks.map((track) => mapTrackOwner(track, ownerNamesById));
-}
-
-function getTrackArtistName(track: Track): string | undefined {
-  return track.tags.artist ?? track.tags.albumArtist ?? track.tags.artists?.[0];
-}
-
-type ArtistMetadata = {
-  name: string;
-  photo?: {
-    path: string;
-  };
-};
-
-type AlbumMetadata = {
-  id?: string;
-  name: string;
-  cover?: {
-    path: string;
-  };
-  tracks?: Track[];
-};
-
-type ResolvedAlbumMetadata = {
-  albumDir: string;
-  id?: string;
-  name?: string;
-  coverPath?: string;
-};
-
-async function resolveArtistPhotoPath(
-  track: Track,
-  cache: Map<string, string | undefined>
-): Promise<string | undefined> {
-  try {
-    const trackAbsolutePath = resolveDataRelativePath(track.path);
-    const albumDir = path.dirname(trackAbsolutePath);
-    const artistDir = path.dirname(albumDir);
-
-    if (cache.has(artistDir)) {
-      return cache.get(artistDir);
-    }
-
-    const metadataPath = path.join(artistDir, ARTIST_METADATA_FILE);
-    const metadata = await readJsonFile<ArtistMetadata | null>(metadataPath, null);
-    const photoPath = metadata?.photo?.path;
-
-    if (!photoPath) {
-      cache.set(artistDir, undefined);
-      return undefined;
-    }
-
-    const absolutePhotoPath = resolveDataRelativePath(photoPath);
-    const hasPhoto = await fileExists(absolutePhotoPath);
-    const resolved = hasPhoto ? photoPath : undefined;
-    cache.set(artistDir, resolved);
-    return resolved;
-  } catch {
-    return undefined;
-  }
-}
-
-async function resolveAlbumMetadataForTrack(
-  track: Track,
-  cache: Map<string, ResolvedAlbumMetadata | undefined>
-): Promise<ResolvedAlbumMetadata | undefined> {
-  try {
-    const trackAbsolutePath = resolveDataRelativePath(track.path);
-    const albumDir = path.dirname(trackAbsolutePath);
-
-    if (cache.has(albumDir)) {
-      return cache.get(albumDir);
-    }
-
-    const metadataPath = path.join(albumDir, ALBUM_METADATA_FILE);
-    const metadata = await readJsonFile<AlbumMetadata | null>(metadataPath, null);
-    const coverPath = metadata?.cover?.path;
-    let resolvedCoverPath: string | undefined;
-
-    if (coverPath) {
-      const absoluteCoverPath = resolveDataRelativePath(coverPath);
-      const hasCover = await fileExists(absoluteCoverPath);
-      resolvedCoverPath = hasCover ? coverPath : undefined;
-    }
-
-    const resolved: ResolvedAlbumMetadata = {
-      albumDir,
-      id: metadata?.id?.trim() ? metadata.id : undefined,
-      name: metadata?.name?.trim() ? metadata.name : undefined,
-      coverPath: resolvedCoverPath
-    };
-
-    cache.set(albumDir, resolved);
-    return resolved;
-  } catch {
-    return undefined;
-  }
-}
-
-async function attachArtistPhotos(
-  tracks: Track[]
-): Promise<Array<{ name: string; trackCount: number; photo?: string; previewTrackId?: string }>> {
-  const base = listArtists(tracks);
-  const photoByArtist = new Map<string, string>();
-  const previewTrackIdByArtist = new Map<string, string>();
-  const cache = new Map<string, string | undefined>();
-
-  for (const track of tracks) {
-    const artistName = getTrackArtistName(track)?.trim();
-    if (!artistName) {
-      continue;
-    }
-
-    if (!previewTrackIdByArtist.has(artistName)) {
-      previewTrackIdByArtist.set(artistName, track.id);
-    }
-
-    if (photoByArtist.has(artistName)) {
-      continue;
-    }
-
-    const photoPath = await resolveArtistPhotoPath(track, cache);
-    if (photoPath) {
-      photoByArtist.set(artistName, photoPath);
-    }
-  }
-
-  return base.map((artist) => ({
-    ...artist,
-    photo: photoByArtist.get(artist.name),
-    previewTrackId: previewTrackIdByArtist.get(artist.name)
-  }));
-}
-
-function normalizeAlbumName(value?: string): string {
-  return (value ?? "")
-    .trim()
-    .toLowerCase()
-    .replace(/\s+/g, " ");
-}
-
-function isCollaborativeAlbumTrack(track: Track): boolean {
-  return Array.isArray(track.tags.artists) && track.tags.artists.length > 1;
-}
-
-function createCollaborativeAlbumId(owner: string, albumName: string): string {
-  return `collab:${owner}:${normalizeAlbumName(albumName)}`;
-}
-
-function parseCollaborativeAlbumId(albumId: string): { owner: string; normalizedAlbumName: string } | null {
-  if (!albumId.startsWith("collab:")) {
-    return null;
-  }
-
-  const parts = albumId.split(":");
-  if (parts.length < 3) {
-    return null;
-  }
-
-  const owner = parts[1]?.trim();
-  const normalizedAlbumName = parts.slice(2).join(":").trim();
-  if (!owner || !normalizedAlbumName) {
-    return null;
-  }
-
-  return { owner, normalizedAlbumName };
-}
-
-async function attachCollaborativeAlbumCovers(
-  tracks: Track[]
-): Promise<Array<{ id?: string; name: string; artist?: string; artists?: string[]; trackCount: number; cover?: string; previewTrackId?: string }>> {
-  const grouped = new Map<
-    string,
-    {
-      id?: string;
-      name: string;
-      artists: Set<string>;
-      trackCount: number;
-      cover?: string;
-      previewTrackId?: string;
-    }
-  >();
-
-  const metadataCache = new Map<string, ResolvedAlbumMetadata | undefined>();
-
-  for (const track of tracks) {
-    const metadata = await resolveAlbumMetadataForTrack(track, metadataCache);
-    const albumName = track.tags.album?.trim() ?? metadata?.name?.trim();
-    if (!albumName) {
-      continue;
-    }
-
-    const collaborative = Boolean(track.tags.album?.trim()) && isCollaborativeAlbumTrack(track);
-    const albumKey = collaborative
-      ? createCollaborativeAlbumId(track.owner, albumName)
-      : metadata?.id ?? `${track.owner}:${metadata?.albumDir ?? normalizeAlbumName(albumName)}`;
-    const current = grouped.get(albumKey);
-    const artistName = getTrackArtistName(track)?.trim();
-    const displayName = metadata?.name?.trim() ? metadata.name : albumName;
-
-    if (!current) {
-      grouped.set(albumKey, {
-        id: collaborative ? albumKey : metadata?.id,
-        name: displayName,
-        artists: artistName ? new Set([artistName]) : new Set<string>(),
-        trackCount: 1,
-        cover: metadata?.coverPath,
-        previewTrackId: track.id
-      });
-      continue;
-    }
-
-    current.trackCount += 1;
-    if (artistName) {
-      current.artists.add(artistName);
-    }
-    if (!current.cover && metadata?.coverPath) {
-      current.cover = metadata.coverPath;
-    }
-    if (!current.previewTrackId) {
-      current.previewTrackId = track.id;
-    }
-  }
-
-  const entries: Array<{ id?: string; name: string; artist?: string; artists?: string[]; trackCount: number; cover?: string; previewTrackId?: string }> = [];
-
-  for (const groupedAlbum of grouped.values()) {
-    const artists = Array.from(groupedAlbum.artists).sort((a, b) => a.localeCompare(b));
-
-    entries.push({
-      id: groupedAlbum.id,
-      name: groupedAlbum.name,
-      artist: artists.length > 0 ? artists.join(", ") : undefined,
-      artists: artists.length > 0 ? artists : undefined,
-      trackCount: groupedAlbum.trackCount,
-      cover: groupedAlbum.cover,
-      previewTrackId: groupedAlbum.previewTrackId
-    });
-  }
-
-  return entries.sort((a, b) => a.name.localeCompare(b.name));
-}
-
-function compareAlbumTrackOrder(a: Track, b: Track): number {
-  const aDisc = a.tags.discNumber ?? Number.MAX_SAFE_INTEGER;
-  const bDisc = b.tags.discNumber ?? Number.MAX_SAFE_INTEGER;
-  if (aDisc !== bDisc) {
-    return aDisc - bDisc;
-  }
-
-  const aTrack = a.tags.trackNumber ?? Number.MAX_SAFE_INTEGER;
-  const bTrack = b.tags.trackNumber ?? Number.MAX_SAFE_INTEGER;
-  if (aTrack !== bTrack) {
-    return aTrack - bTrack;
-  }
-
-  const aTitle = a.tags.title ?? a.path;
-  const bTitle = b.tags.title ?? b.path;
-  return aTitle.localeCompare(bTitle);
 }
 
 export function createLibraryRouter(indexStore: IndexStore): Router {
